Add note about GameMaker handling booleans as 0 and 1
Due to backwards compatibility, boolean values in GameMaker are internally stored as `real`s (0 or 1). While GameMaker does have an internal boolean data type, it's rarely used -- the vast majority of functions that return "booleans" actually return 0 or 1. This isn't a problem when you're working inside of GameMaker. But when it comes to transfering data between programming languages, as is the case here, this is a huge burden. If those values are passed as arguments to a function, they will, for the reason stated above, become `float`s in Python (0.0 and 1.0), which aren't even accepted as booleans by most functions! The only way around this is to call `bool()` on every boolean value you pass as an argument, as it is one of the few functions that return the **true** boolean type. This way, the boolean values will be correctly received by Python as `True` and `False`.